+++ spec/dummy/demo_app/app/controllers/application_controller.rb
@@ -2,6 +2,9 @@
 
 module DemoApp
   class ApplicationController < Charming::Controller
+    include Charming::Shell::Sidebar
+    include Charming::Shell::Palette
+
     layout Layouts::ApplicationLayout
     focus_ring :sidebar, :content